requirement.txt
pandas==2.0.3
google-generativeai==0.7.0
chardet==5.1.0

